fix - [teams edit] alerte suppression equipe inscrite à un événement.

+++ front/src/app/modules/client/pages/teams/teams.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Team } from 'src/app/models/teams/team';
+import { EventService } from 'src/app/modules/ms-api/event/event.service';
 import { TeamService } from 'src/app/modules/ms-api/team/team.service';
 import { AddTeamComponent } from 'src/app/modules/ms-ui/components/add-team/add-team.component';
 
@@ -15,9 +16,11 @@ export class TeamsComponent implements OnInit {
   teams : Team[];
   id: string;
   displayEdition = false;
+  displayDeletionTeamError = false;
 
   constructor(
     private teamService: TeamService,
+    private eventService: EventService,
     private route : ActivatedRoute,
     private router: Router,
     private modalService : NgbModal
@@ -37,9 +40,18 @@ export class TeamsComponent implements OnInit {
   }
 
   deleteTeam(teamId : string) {
-    this.teamService.deleteTeam(teamId).subscribe(() => {
-      this.initTeams();
-      this.router.navigateByUrl(`/captain/${this.id}/teams`)
+    this.eventService.findByTeamRegistration(teamId).subscribe(data => {
+      if(data.length != 0) {
+        this.displayDeletionTeamError = true;
+        setTimeout(() => {
+          this.displayDeletionTeamError = false;
+        }, 7000)
+      } else {
+        this.teamService.deleteTeam(teamId).subscribe(() => {
+          this.initTeams();
+          this.router.navigateByUrl(`/captain/${this.id}/teams`)
+        })
+      }
     })
   }
